+++ src/envelope_widget.cpp
@@ -29,6 +29,7 @@
 #include "ViewState.h"
 #include "GeonkickModel.h"
 #include "OscillatorModel.h"
+#include "GeonkickConfig.h"
 
 #include "RkContainer.h"
 #include "RkLabel.h"
@@ -40,6 +41,9 @@ RK_DECLARE_IMAGE_RC(layer2_env);
 RK_DECLARE_IMAGE_RC(layer2_env_active);
 RK_DECLARE_IMAGE_RC(layer3_env);
 RK_DECLARE_IMAGE_RC(layer3_env_active);
+RK_DECLARE_IMAGE_RC(bezier_mode_button);
+RK_DECLARE_IMAGE_RC(bezier_mode_button_on);
+RK_DECLARE_IMAGE_RC(bezier_mode_button_hover);
 
 EnvelopeWidget::EnvelopeWidget(GeonkickWidget *parent, GeonkickModel *model)
         : GeonkickWidget(parent)
@@ -49,6 +53,7 @@ EnvelopeWidget::EnvelopeWidget(GeonkickWidget *parent, GeonkickModel *model)
         , layer2Button{nullptr}
         , layer3Button{nullptr}
 #endif // GEONKICK_BASIC_VERSION
+        , bezierModeButton {nullptr}
         , geonkickModel{model}
         , dspProxy{geonkickModel->getDspProxy()}
         , oscillators{geonkickModel->getOscillatorModels()}
@@ -83,9 +88,7 @@ EnvelopeWidget::EnvelopeWidget(GeonkickWidget *parent, GeonkickModel *model)
         generalEnvelope->setCategory(Envelope::Category::InstrumentGlobal);
         envelopes.insert({static_cast<int>(Envelope::Category::InstrumentGlobal),
                         std::move(generalEnvelope)});
-#ifndef GEONKICK_BASIC_VERSION
         createButtomMenu();
-#endif // GEONKICK_BASIC_VERSION
         showEnvelope(Envelope::Category::InstrumentGlobal, Envelope::Type::Amplitude);
         RK_ACT_BIND(viewState(), envelopeChanged,
                     RK_ACT_ARGS(Envelope::Category category, Envelope::Type envelope),
@@ -96,25 +99,54 @@ EnvelopeWidget::EnvelopeWidget(GeonkickWidget *parent, GeonkickModel *model)
         drawArea->show();
 }
 
-#ifndef GEONKICK_BASIC_VERSION
 void EnvelopeWidget::createButtomMenu()
 {
         auto buttomAreaWidget = new GeonkickWidget(drawArea);
         buttomAreaWidget->setBackgroundColor(40, 40, 40);
-        buttomAreaWidget->setFixedSize(90, 20);
+#ifndef GEONKICK_BASIC_VERSION
+        buttomAreaWidget->setFixedSize(120, 20);
+#else
+        buttomAreaWidget->setFixedSize(25, 20);
+#endif  // GEONKICK_BASIC_VERSION
         buttomAreaWidget->setPosition(55 + drawArea->x(),
                                       drawArea->y() + drawArea->height() - buttomAreaWidget->height() - 6);
 
         auto menuContainer = new RkContainer(buttomAreaWidget);
+#ifndef GEONKICK_BASIC_VERSION
         createLayersButtons(buttomAreaWidget);
         menuContainer->addWidget(layer1Button);
         menuContainer->addSpace(5);
         menuContainer->addWidget(layer2Button);
         menuContainer->addSpace(5);
         menuContainer->addWidget(layer3Button);
+        menuContainer->addSpace(10);
+#endif // GEONKICK_BASIC_VERSION
+        createBezierModeControls(buttomAreaWidget, menuContainer);
         buttomAreaWidget->show();
 }
-#endif // GEONKICK_BASIC_VERSION
+
+void EnvelopeWidget::createBezierModeControls(GeonkickWidget* widget, RkContainer *container)
+{
+        bezierModeButton = new GeonkickButton(widget);
+        bezierModeButton->setSize(24, 18);
+        bezierModeButton->setCheckable(true);
+        bezierModeButton->setPressed(GeonkickConfig().isBezierMode());
+        bezierModeButton->setImage(RK_RC_IMAGE(bezier_mode_button),
+                                  RkButton::State::Unpressed);
+        bezierModeButton->setImage(RK_RC_IMAGE(bezier_mode_button_on),
+                                  RkButton::State::Pressed);
+        bezierModeButton->setImage(RK_RC_IMAGE(bezier_mode_button_hover),
+                                  RkButton::State::UnpressedHover);
+        bezierModeButton->setImage(RK_RC_IMAGE(bezier_mode_button_hover),
+                                  RkButton::State::PressedHover);
+        RK_ACT_BIND(bezierModeButton,
+                    toggled,
+                    RK_ACT_ARGS(bool pressed),
+                    this,
+                    enableBezierMode(pressed));
+
+        container->addWidget(bezierModeButton);
+}
 
 void EnvelopeWidget::createPointInfoLabel()
 {
@@ -241,4 +273,10 @@ OscillatorModel* EnvelopeWidget::getCurrentOscillator() const
 void EnvelopeWidget::setPointEditingMode(bool b)
 {
         drawArea->setPointEditingMode(b);
+}
+
+void EnvelopeWidget::enableBezierMode(bool b)
+{
+        GeonkickConfig().setBezierMode(b);
+        drawArea->setBezierMode(b);
 }